% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spe.R
\name{allocateRegion}
\alias{allocateRegion}
\title{Allocate points to annotated regions}
\usage{
allocateRegion(x, y, regions, ...)
}
\arguments{
\item{x}{a numeric, containing x-coordinates}

\item{y}{a numeric, containing y-coordinates}

\item{regions}{a character, specifying the regions}
}
\value{
a numeric, containing the index of bins to which each point was allocated
}
\description{
Allocate points to annotated regions
}
