% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spe.R
\name{allocateSquare}
\alias{allocateSquare}
\title{Allocate points to a regular square grid}
\usage{
allocateSquare(x, y, bins = 30, ...)
}
\arguments{
\item{x}{a numeric, containing x-coordinates}

\item{y}{a numeric, containing y-coordinates}

\item{bins}{a numeric, stating the number of bins partitioning the range of x}
}
\value{
a numeric, containing the index of bins to which each point was allocated
}
\description{
Allocate points to a regular square grid
}
