% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spe.R
\name{tx2spe}
\alias{tx2spe}
\title{Compute cell/bin-level summaries and save to a SpatialExperiment object}
\usage{
tx2spe(x, bin = c("cell", "hex", "square", "region"), nbins = 100)
}
\arguments{
\item{x}{a data.frame containing a sub-cellular localised dataset from the SubcellularSpatialData package.}

\item{bin}{a character, stating whether transcripts should be binned into cells or hexbins. The default is 'cell', however, it is advisable to use 'hexbin' when cell identification is not accurate (e.g., when cell boundaries are inferred only from nuclei and not directly measured using cytoplasmic/cell membrane stains).}

\item{nbins}{a numeric, stating the number of bins to create in the x and y axes}
}
\value{
a SpatialExperiment object containing cell/hexbin-level expression quantification.
}
\description{
This function summarises transcript detection (or DNA nanoball spot counts) into cells or hexbins and returns a SpatialExperiment object. Annotations of cells or hexbins are summarised such that the most frequent annotation is used.
}
