% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubcellularSpatialData.R
\docType{data}
\name{tx_small}
\alias{tx_small}
\title{Sample annotated sub-cellular localised spatial 'omics dataset}
\format{
A data.frame where each row represents a transcript detection. The full description of these transcript tables is available in the vignettes or in the documentation for the \code{tx2spe()} function.
}
\usage{
tx_small
}
\description{
This object stores a sample transcript table to demonstrate the format of data stored within this package. The sample dataset contains 100,000 transcript detections from 2 breast cancer samples profiled using the 10x Xenium platform.
}
\keyword{datasets}
